split domains from network
