+++ packages/auto-utils/src/constants/index.ts
@@ -1,5 +1,6 @@
 // file: src/constants/index.ts
 
+export * from './domain'
 export * from './network'
 export * from './token'
 export * from './wallet'
